+++ src/egeo-demo/public_api.ts
@@ -15,6 +15,7 @@ export { StCheckboxDemoModule } from './st-checkbox-demo/st-checkbox-demo.module
 export { StDropdownDemoModule } from './st-dropdown-demo/st-dropdown-demo.module';
 export { StDropdownMenuDemoModule } from './st-dropdown-menu-demo/st-dropdown-menu-demo.module';
 export { StFormDemoModule } from './st-form-demo/st-form-demo.module';
+export { StHeaderDemoModule } from './st-header-demo/st-header-demo.module';
 export { StHelpDemoModule } from './st-help-demo/st-help-demo.module';
 export { StInfoBoxDemoModule } from './st-info-box-demo/st-info-box-demo.module';
 export { StInputDemoModule } from './st-input-demo/st-input-demo.module';